requirements.txt
python-twitter >= 3.3.0
b4msa >= 1.2.1
jsonpickle >= 0.9.5